requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==2.0.20
pydantic==1.10.12
python-jose==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.1.0
python-multipart==0.0.6
requests==2.31.0
